pyproject.toml
[project]
name = "reporadar"
version = "0.1.0"
description = "Discover similar GitHub repositories via dual-vector semantic search"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "httpx>=0.27",
    "qdrant-client>=1.8",
    "sentence-transformers>=2.5",
    "pydantic-settings>=2.1",
    "python-jose[cryptography]>=3.3",
    "beautifulsoup4>=4.12",
    "pyyaml>=6.0",
    "slowapi>=0.1.9",
    "tqdm>=4.66",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.3",
    "respx>=0.20",
]

[tool.setuptools.packages.find]
include = ["app*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = ["slow: marks tests as slow (loads ML model)"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
